pyproject.toml
[tool.poetry]
name = "weather-forecasting"
version = "0.1.0"
description = ""
authors = ["atabak-hooshangi <atabak.hooshangi.1990@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
cassandra-driver = "^3.29.1"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.5"
certifi = "^2024.2.2"
pytz = "^2024.1"
pandas = "^2.2.2"
dask = {extras = ["dataframe"], version = "^2025.4.0"}
aioredis = "^2.0.1"
redis = "^5.0.8"
aiofiles = "^24.1.0"
pyarrow = "^19.0.1"
pyproj = "^3.7.1"
meteostat = "^1.6.8"
requests = "^2.32.3"
tabulate = "^0.9.0"
scikit-learn = "^1.6.1"
lightgbm = "^4.6.0"
matplotlib = "^3.10.1"
jupyterlab = "^4.4.1"
notebook = "^7.4.1"
torch = "^2.7.0"
pytorch-lightning = "^2.5.1"
pytorch-forecasting = "^1.3.0"
numpy = ">=1.26,<2.2"
tensorflow = "^2.19.0"
seaborn = "^0.13.2"
fastapi = "^0.115.12"
pydantic-settings = "^2.9.1"
itsdangerous = "^2.2.0"
uvicorn = "^0.34.2"
fastapi-cache2 = "^0.2.2"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false